Fixes issue with missing functions in outline.
This was due to the fact that some `locals.scm` files contain multiple
definitions within a single query. This introduces some ambiguities in
the mapping from definitions to their containing scopes and thus we did
not think it was possible. WELL, it is possible. In fact, it occurs in
lua's locals file.  We fixed this in two ways:

1. Split multiple definitions into their own TSOutline object, although
the scope only gets linked with the first definition (although the
definitions still keep their link to the scope).

2. In the case where `definition.associated` is one of two definitions,
we simply use the other definition while maintaining a reference to
definition.associated for display purposes.